Rename all source files to '*.gpp.md'
Need to differenciate from README.md and possibly other Markdown files
that might get added in the future.
Also, these files are meant to be pre-processed specifically with GPP,
so it makes sense adding that to their name.
Keep the '.md' final extension so editors interpret them as Markdown
files (as opposed to eg. '*.md.gpp' or '*.md.in')